Move client tests to own files
